pyproject.toml
[tool.poetry]
name = "db-huggy"
version = "0.1.0"
description = ""
authors = ["AngeloGagno <angelogagno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
aiohttp = "^3.11.11"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
